Rename EnrollmentActions component to StickyActions
