+++ frontend/src/components/templates/Form/SubmissionPanel/index.tsx
@@ -5,7 +5,7 @@ import {
 } from '../../../../config/event-configuration';
 import { processEvent } from '../../../../lib/process-event';
 import Loader from '../../../atoms/Loader';
-import EnrollmentActions from '../../../molecules/EnrollmentActions';
+import StickyActions from '../../../molecules/StickyActions';
 import ConfirmationModal from '../../../organisms/ConfirmationModal';
 import { useFormSubmission } from './hooks/use-form-submission';
 import Prompt from './Prompt';
@@ -36,7 +36,7 @@ const SubmissionPanel: FunctionComponent<Props> = ({
 
   return (
     <>
-      <EnrollmentActions
+      <StickyActions
         disabled={!!pendingEvent || loading}
         acl={enrollment.acl}
         onEventButtonClick={onEventButtonClick}
